refactor(payment): Phase 4 split EfPaymentStore from PaymentContext
- Introduce EfPaymentStore implementing IPaymentStore by re-issuing
  against ctx.Payments/ctx.OrderCustomers; ExecuteAsync delegates to
  PaymentContext.ExecuteAsync (Translate switch + outbox UoW handoff
  removed/rerouted in Phase 6).
- Flip DI: services.AddScoped<IPaymentStore, EfPaymentStore>(); strip
  IPaymentStore method bodies and interface from PaymentContext.
