+++ payment-microservice/Payment.Service/Infrastructure/Data/EntityFramework/EntityFrameworkExtensions.cs
@@ -17,6 +17,6 @@ public static class EntityFrameworkExtensions
                         errorNumbersToAdd: [0]);
                 }));
 
-        services.AddScoped<IPaymentStore, PaymentContext>();
+        services.AddScoped<IPaymentStore, EfPaymentStore>();
     }
 }
